requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=4.1.0
lxml>=4.9.0
reagentic
aiogram
python-dotenv 